+++ lib/testing.dart
@@ -2,40 +2,26 @@ library bark.testing;
 
 import 'package:bark/src/bloc.dart';
 
-class TestComponent<T> implements BlocState {
+class TestComponent<T> implements ComponentGlue {
   final Bloc<T> bloc;
   final List<BlocMessage> messages = [];
 
-  T Function<T>(BlocMessage<T>) mockMessageHandler;
-
-  BlocContext _context;
-  T _latestValue;
+  T Function<T>(BlocMessage<T>)? mockMessageHandler;
 
   TestComponent(this.bloc) {
-    _context = BlocContext(this);
+    bloc.attachComponent(this);
   }
 
-  T get latestValue => _latestValue ??= _fetchValue();
-
-  T _fetchValue() {
-    final value = bloc.viewModel(_context);
-    _context.clearUntracked();
-    return value;
-  }
+  T get latestValue => bloc.value;
 
   @override
   T handleMessage<T>(BlocMessage<T> message) {
     messages.add(message);
 
     if (mockMessageHandler != null) {
-      return mockMessageHandler(message);
+      return mockMessageHandler!(message);
     }
 
     return message.onDeliveryFailure();
-  }
-
-  @override
-  void onTrackerUpdated(Tracker tracker) {
-    _latestValue = null;
   }
 }